@@ -13,53 +13,60 @@
     "postinstall": "nuxt prepare"
   },
   "devDependencies": {
-    "@iconify-json/fa6-brands": "^1.1.13",
-    "@iconify-json/fa6-solid": "^1.1.15",
-    "@iconify-json/ic": "^1.1.14",
-    "@iconify-json/ion": "^1.1.4",
-    "autoprefixer": "^10.4.8",
-    "css-loader": "^6.7.1",
-    "cssnano": "^5.1.13",
-    "mini-css-extract-plugin": "^2.6.1",
-    "nuxt": "3.0.0-rc.11",
+    "@iconify-json/fa6-brands": "^1.2.5",
+    "@iconify-json/fa6-solid": "^1.2.3",
+    "@iconify-json/ic": "^1.2.2",
+    "@iconify-json/ion": "^1.2.4",
+    "@tailwindcss/postcss": "^4.1.11",
+    "@types/node": "^24.0.13",
+    "css-loader": "^7.1.2",
+    "cssnano": "^7.0.7",
+    "mini-css-extract-plugin": "^2.9.2",
+    "nuxt": "3.17.7",
     "nuxt-compress": "^5.0.0",
-    "postcss": "^8.4.16",
-    "postcss-loader": "^7.0.1",
-    "sass-loader": "^13.3.2",
-    "unplugin-icons": "^0.17.0"
+    "postcss": "^8.5.6",
+    "postcss-loader": "^8.1.1",
+    "sass-loader": "^16.0.5",
+    "tailwindcss": "^4.1.11",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3",
+    "unplugin-icons": "^22.1.0",
+    "webpack": "^5.100.1"
   },
   "dependencies": {
-    "@emailjs/browser": "^3.7.0",
-    "@fortawesome/fontawesome-svg-core": "^6.2.0",
-    "@fortawesome/free-brands-svg-icons": "^6.2.0",
-    "@fortawesome/free-regular-svg-icons": "^6.2.0",
-    "@fortawesome/free-solid-svg-icons": "^6.2.0",
-    "@fortawesome/vue-fontawesome": "^3.0.1",
-    "@nuxtjs/color-mode": "^3.1.6",
-    "@nuxtjs/dotenv": "^1.4.1",
-    "@nuxtjs/tailwindcss": "^5.3.3",
-    "@pinia/nuxt": "^0.4.2",
-    "@tailwindcss/typography": "^0.5.7",
-    "@vueuse/core": "^9.3.0",
-    "bootstrap-vue": "^2.22.0",
-    "cropperjs": "^1.6.1",
-    "dayjs": "^1.11.10",
-    "highlight.js": "^11.8.0",
-    "jquery": "^3.6.1",
-    "katex": "^0.16.9",
-    "md-editor-v3": "^4.7.0",
-    "mermaid": "^10.5.0",
-    "node-fetch": "^3.2.10",
-    "pinia": "^2.0.22",
-    "pinia-plugin-persistedstate": "^2.2.0",
-    "prettier": "^3.0.3",
-    "sanitize-html": "^2.11.0",
-    "sass": "^1.69.0",
+    "@emailjs/browser": "^4.4.1",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/vue-fontawesome": "^3.0.8",
+    "@nuxt/content": "^3.6.3",
+    "@nuxtjs/color-mode": "^3.5.2",
+    "@nuxtjs/dotenv": "^1.4.2",
+    "@nuxtjs/gtm": "^2.4.0",
+    "@nuxtjs/sitemap": "^7.4.3",
+    "@nuxtjs/tailwindcss": "^6.14.0",
+    "@pinia/nuxt": "^0.11.1",
+    "@tailwindcss/typography": "^0.5.16",
+    "@vueuse/core": "^13.5.0",
+    "better-sqlite3": "^12.2.0",
+    "bootstrap-vue": "^2.23.1",
+    "cropperjs": "^2.0.0",
+    "dayjs": "^1.11.13",
+    "highlight.js": "^11.11.1",
+    "jquery": "^3.7.1",
+    "katex": "^0.16.22",
+    "md-editor-v3": "^5.8.0",
+    "mermaid": "^11.8.1",
+    "node-fetch": "^3.3.2",
+    "pinia": "^3.0.3",
+    "pinia-plugin-persistedstate": "^4.4.1",
+    "prettier": "^3.6.2",
+    "sanitize-html": "^2.17.0",
+    "sass": "^1.89.2",
     "screenfull": "^6.0.2",
-    "sweetalert2": "^11.4.38",
-    "tailwindcss": "^3.1.8",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.74.0"
+    "sweetalert2": "^11.22.2",
+    "url-loader": "^4.1.1"
   },
-  "packageManager": "yarn@4.3.1"
+  "packageManager": "yarn@4.9.2"
 }
