@@ -1,7 +1,7 @@
 <template>
     <main class="draftPosts">
         <logo/>
-        <body >
+        <section >
             <div v-if="posts">
                 <div class="submit">
                     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/solid.css">
@@ -59,7 +59,7 @@
             <div class="card" v-else>
                 <img  src="@/assets/loading.gif" alt="loading"/>
             </div>
-        </body>
+        </section>
 
     </main>
 </template>
@@ -201,7 +201,7 @@ const deletePost = (id)=>{
     margin-top: 30%;
   }
 
-body{
+section{
     margin-left: 50px;
     margin-top: 120px;
     background-color: #f1f5f9;
