@@ -1,1 +1,1 @@
-export const API = 'https://meshneerstackback.neerstack.com'
+export const API = process.env.NUXT_API || 'https://codinmage-server.onrender.com'
